bridge/requirements.txt
cffi==1.17.1
cryptography==45.0.5
pycparser==2.22
python-dotenv==1.1.1
pyzmq==27.0.0
redis==6.2.0
tenacity==9.1.2
